+++ src/WS/Message/SendMessage.ts
@@ -1,19 +1,18 @@
 import { storeClientConfig } from "~/stores/ClientConfig.ts";
 import { setStoreHasNewMessage } from "~/stores/HasNewMessage.ts";
-import { addMessage } from "~/stores/History.ts";
+import { addMessage, storeHistory } from "~/stores/History.ts";
 import { storeMyUserinfo } from "~/stores/MyUserinfo.ts";
 import { setStoreMessageReadTime } from "~/stores/Readtime";
 import type { IMessage } from "~/types/Message.ts";
 import { notifyIt } from "~/utils/Notify.ts";
+import UpdateReadTimeOnRemoteAndStore from "~/utils/UpdateReadTimeOnRemoteAndStore.util";
 
 export default function WSSendMessage(dat: IMessage) {
   //console.log("WSSendMessage :: triggered dat->", dat);
 
-  //履歴に追加
-  addMessage(dat);
-
   //もし受け取ったメッセージのチャンネルにいない、あるいはフォーカスしていないなら新着設定
   if ((!location.pathname.includes(dat.channelId) || !document.hasFocus()) && storeMyUserinfo.id !== dat.userId) {
+    //console.log("SendMessage :: WSSendMessage : 新着登録");
     setStoreHasNewMessage((hnm) => {
       return {
         ...hnm,
@@ -26,16 +25,11 @@ export default function WSSendMessage(dat: IMessage) {
       //console.log("WSSendMessage :: notifyIt ->", dat.userId, dat.content);
       notifyIt(dat.userId, dat.content, { channelId: dat.channelId });
     }
-  } else { //普通にアクティブなら時差表示用既読時間を更新
-    setStoreMessageReadTime((prev) => {
-      const newReadTime = {
-        channelId: dat.channelId,
-        readTime: dat.createdAt,
-        readTimeBefore: dat.createdAt
-      };
-      const newStore = prev.filter((c) => c.channelId !== dat.channelId);
-      newStore.push({ ...newReadTime });
-      return newStore;
-    });
+  } else if (storeHistory[dat.channelId]?.atEnd || storeMyUserinfo.id === dat.userId) { //それ以外で履歴末端まで行ってるなら既読時間更新
+    const updateReadtimeBeforeToo = storeMyUserinfo.id === dat.userId || document.hasFocus();
+    UpdateReadTimeOnRemoteAndStore(dat.channelId, dat.createdAt, { copyToReadTimeBefore: updateReadtimeBeforeToo });
   }
+
+  //履歴に追加
+  addMessage(dat);
 }
